+++ src/markets/perps/interfaces/ISettlementModule.sol
@@ -13,7 +13,7 @@ import { SD59x18 } from "@prb-math/SD59x18.sol";
 interface ISettlementModule {
     event LogSettleOrder(
         address indexed sender,
-        uint256 indexed accountId,
+        uint128 indexed accountId,
         uint128 indexed marketId,
         int256 pnl,
         Position.Data newPosition
